+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/DeleteHubResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
 import java.math.BigDecimal;
