+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/HubDetailResponseDto.java
@@ -1,6 +1,7 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
 import java.math.BigDecimal;
+import java.util.UUID;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -8,11 +9,16 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class CreateHubRequestDto {
+public class HubDetailResponseDto {
 
   private long userId;
   private String hubName;
   private String address;
   private BigDecimal latitude;
   private BigDecimal longitude;
+
+  // 테스트를 위한 임시 추가
+  private long hubId;
+
+
 }
