+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/UpdateHubRequestDto.java
@@ -1,18 +1,19 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
 import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
-@Builder
 @Getter
-public class UpdateHubResponseDto {
+@Setter
+@Builder
+public class UpdateHubRequestDto {
   private long userId;
   private String hubName;
   private String address;
   private BigDecimal latitude;
   private BigDecimal longitude;
 
-  // 테스트를 위한 임시 추가
-  private long hubId;
+
 }
