+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/application/dto/UpdateHubResponseDto.java
@@ -1,19 +1,18 @@
-package com.sparta.logistics.hub_service.dto;
+package com.sparta.logistics.hub_service.application.dto;
 
 import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
-@Getter
-@Setter
 @Builder
-public class UpdateHubRequestDto {
+@Getter
+public class UpdateHubResponseDto {
   private long userId;
   private String hubName;
   private String address;
   private BigDecimal latitude;
   private BigDecimal longitude;
 
-
+  // 테스트를 위한 임시 추가
+  private long hubId;
 }
